ogre helpers point_cloud: add size into bounds
Add the size of the point visual into the bounding box.
For small or zero sized points this makes little difference,
but for large points (such as visualizing an octomap with
large compressed regions), it is important to add the bounds
to keep the points visible at the very edges of the cloud.